fix(ppu): Correct tile address calculation in fetcher
This commit fixes a panic in the PPU fetcher caused by an incorrect address calculation for the `0x8800` tile data addressing mode. The previous implementation could cause an out-of-bounds memory access when converting the signed tile index, leading to a panic.

The fix ensures the calculation is done using `i32` to prevent overflow and incorrect sign extension, which resolves the panic and allows the boot ROM to render correctly.

This also includes the original PPU feature implementations:
- Implemented writing to the LYC register (0xFF45).
- Implemented background scrolling (SCX and SCY).
- Implemented the window display feature.
- Refactored the PPU Fetcher to handle both tile data addressing modes (0x8000 and 0x8800) based on LCDC.4.
- Fixed the STAT register interrupt logic to correctly trigger interrupts for all modes (H-Blank, V-Blank, OAM-Scan) and for the LYC=LY condition.

+++ src/ppu/fetcher.rs
@@ -142,7 +142,7 @@ impl Fetcher {
         let offset = if self.lcdc.tile_data_select() {
             0x8000 + (self.tile_id as u16 * 16)
         } else {
-            0x9000 + ((self.tile_id as i8 as u16) * 16)
+            (0x9000 as i32 + (self.tile_id as i8 as i32 * 16)) as u16
         };
 
         // Then, from that starting offset, we compute the final address to read
